# create KernelMinMax to optimize the performance of histogram op in GPU