<!-- Demo: https://github.com/PaddlePaddle/Paddle/pull/24810 -->
### PR types
<!-- One of [ New features | Bug fixes | Function optimization | Performance optimization | Breaking changes | Others ] -->
Performance optimization 
### PR changes
<!-- One of [ OPs | APIs | Docs | Others ] -->
 OPs 
### Describe
<!-- Describe what this PR does -->
当前Paddle采用自主编写的CUDA Kernel执行Histogram的核心计算部分，但是在确定直方图边界时使用Eigen进行计算，存在一定的优化空间
设计文档：https://github.com/PaddlePaddle/community/blob/master/rfcs/OPs-Perf/20230328_histogram_op_optimization.md

* 开发环境
   1. 设备：Tesla V100
   2. 环境：CUDA11.2，cuDNN 8

* 优化方法
   *  关键是使用__global__ kernel的方式实现了KernelMinMax，加速Histogram确定直方图边界的计算部分，从而提高Histogram算子在GPU上的计算性能。

完成优化后，Paddle与优化前的Paddle的性能对比效果:
| Case No. | device| input_shape | input_type | bins | min | max |Paddle Perf(ms) |old Paddle Perf(ms) |diff|
|---|---|---|---|---|---|---|---|---|---|
| 1 | Tesla V100 | [16, 64] | int32   | 100 | 0 | 0 | 0.01220 |0.09403|faster than 670.74% |
| 2 | Tesla V100 | [16, 64] | int64   | 100 | 0 | 0 | 0.01515 |0.13624|faster than 799.27%|
| 3 | Tesla V100 | [16, 64] | float32 | 100 | 0 | 0 | 0.01501 |0.01889|faster than 25.85%|

完成优化后，Paddle与Pytorch的性能对比效果如下:
| Case No. | device| input_shape | input_type | bins | min | max |Paddle Perf(ms) |Pytorch Perf(ms) |diff|
|---|---|---|---|---|---|---|---|---|---|
| 1 | Tesla V100 | [16, 64] | int32   | 100 | 0 | 0 | 0.01220 |0.02255|faster than 84.84%|
| 2 | Tesla V100 | [16, 64] | int64   | 100 | 0 | 0 | 0.01515 |0.03424|faster than 126.01%|
| 3 | Tesla V100 | [16, 64] | float32 | 100 | 0 | 0 | 0.01501 |0.02250|faster than 49.90%|

针对三种不同case, 优化后的性能有不同程度的提升。